+++ chpl/chpl-api/src/test/java/gov/healthit/chpl/web/controller/SurveillanceReportControllerQuarterlyReportsTest.java
@@ -1,17 +1,16 @@
 package gov.healthit.chpl.web.controller;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
+import static org.junit.Assert.assertNotNull;
+import static org.junit.Assert.assertNull;
+import static org.junit.Assert.assertTrue;
 
 import java.util.List;
 
-import org.ff4j.FF4j;
-import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Mockito;
-import org.mockito.MockitoAnnotations;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.env.Environment;
 import org.springframework.security.access.AccessDeniedException;
@@ -29,13 +28,11 @@ import com.fasterxml.jackson.core.JsonProcessingException;
 import com.github.springtestdbunit.DbUnitTestExecutionListener;
 import com.github.springtestdbunit.annotation.DatabaseSetup;
 
-import gov.healthit.chpl.FeatureList;
 import gov.healthit.chpl.UnitTestUtil;
 import gov.healthit.chpl.auth.permission.GrantedPermission;
 import gov.healthit.chpl.auth.user.JWTAuthenticatedUser;
 import gov.healthit.chpl.caching.UnitTestRules;
 import gov.healthit.chpl.domain.CertificationBody;
-import gov.healthit.chpl.domain.CertifiedProduct;
 import gov.healthit.chpl.domain.surveillance.report.QuarterlyReport;
 import gov.healthit.chpl.domain.surveillance.report.RelevantListing;
 import gov.healthit.chpl.exception.EntityCreationException;
@@ -60,9 +57,6 @@ public class SurveillanceReportControllerQuarterlyReportsTest {
     @Autowired
     private SurveillanceReportController reportController;
 
-    @Autowired
-    private FF4j ff4j;
-
     @Rule
     @Autowired
     public UnitTestRules cacheInvalidationRule;
@@ -82,12 +76,6 @@ public class SurveillanceReportControllerQuarterlyReportsTest {
         acbUser.setFriendlyName("User3");
         acbUser.setSubjectName("testUser3");
         acbUser.getPermissions().add(new GrantedPermission("ROLE_ACB"));
-    }
-
-    @Before
-    public void setup() {
-        MockitoAnnotations.initMocks(this);
-        Mockito.doReturn(true).when(ff4j).check(FeatureList.SURVEILLANCE_REPORTING);
     }
 
     @Transactional