+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/activity/ActivityMetadataBuilder.java
@@ -103,8 +103,10 @@ public abstract class ActivityMetadataBuilder {
             break;
         case ANNUAL_REPORT:
             metadata = new AnnualReportActivityMetadata();
+            break;
         case CHANGE_REQUEST:
             metadata = new ChangeRequestActivityMetadata();
+            break;
         case API_KEY:
             metadata = new ApiKeyManagementActivityMetadata();
             break;