+++ chpl/chpl-service/src/main/java/gov/healthit/chpl/changerequest/validation/ChangeRequestValidationFactory.java
@@ -57,7 +57,6 @@ public class ChangeRequestValidationFactory {
         this.commentRequiredValidation = commentRequiredValidation;
         this.roleAcbHasMultipleCertificationBodiesValidation = roleAcbHasMultipleCertificationBodiesValidation;
         this.websiteValidation = websiteValidation;
-
     }
 
     public ValidationRule<ChangeRequestValidationContext> getRule(String name) {