+++ RELEASE_NOTES.md
@@ -1,5 +1,22 @@
 # Release Notes
 
+## Version 21.12.0
+_4 November 2019_
+
+### Features
+* Do not require test tools or UCD Processes for 2014 listings that have ICS
+
+---
+
+## Version 21.11.0
+_24 October 2019_
+
+### Features
+* Add Quartz job that will update a set of listing's status
+* Add User scheduled job to generate and send a report of Pending Change Requests
+
+---
+
 ## Version 21.10.0
 _21 October 2019_